Fix bug #21 (#22)
Address issue #21.

When store is performed locally, it was supposed to perform check on the copy of the entry that we are currently
operating on versus the one that after performing the load. The problem was the perform load updated the
current version of the entry and the check was never eval to true.